requirements.txt
# Основные зависимости для AI-агента (совместимые версии)
python-telegram-bot==20.7

# LangChain экосистема (совместимые версии)
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.63
langchain-openai==0.2.8
langgraph==0.2.39
langsmith==0.1.147

# RAG и векторные БД (обновленные совместимые версии)
chromadb==0.5.20
langchain-chroma==0.1.4
faiss-cpu==1.11.0
sentence-transformers==2.6.1

# База данных
SQLAlchemy==2.0.23

# Обработка данных
numpy==1.26.4
scikit-learn==1.6.1

# HTTP клиенты
aiohttp==3.11.18
httpx==0.25.2

# Утилиты
python-dotenv==1.0.0
pydantic==2.11.4
pydantic-settings==2.1.0
PyYAML==6.0.2
tqdm==4.67.1

# Логирование
coloredlogs==15.0.1

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1